Rename runtime.
